BI Analytica
This project is a part of the development of dashboard analytics developed for an IT firm with the help of dash plotly. The firm has a client  which provides the courier services and all the billing, generating reports, and database handling is done by this firm. Now they want to migrate these systems from legacy technologies to new technologies. The part of the project provided to us is to generate and Visualize a dashboard from the given database which will display various parameters useful for business analysis, financial optimization and determining which all places, branches and customers are profitable and have high business trading volume. Also reports will be generated and various views in the dashboard will be provided & also future prediction from all the historical data provided for the next month is done, and these all parameters helps in making an informed business decision by the management.